+++ frontend/src/views/Scetch/classes/ScetchCanvasState.ts
@@ -1,6 +1,5 @@
 import { CanvasPosition } from './CanvasPosition';
 import { Folder } from './ScetchItems/Folder';
-import { Circle } from './ScetchItems/Circle';
 
 export type ScetchCanvasStateOptions = {
   debug: boolean;