+++ frontend/src/views/Scetch/components/ScetchSideMenu/ScetchSideMenu.tsx
@@ -1,0 +1,6 @@
+import { ScetchSideMenuProps } from './types';
+import './styles.scss';
+
+export function ScetchSideMenu(props: ScetchSideMenuProps) {
+  return <div className="ScetchSideMenu">ScetchSideMenu</div>;
+}