+++ uikit/src/components/KButton/components/KButtonIcon/KButtonIcon.tsx
@@ -1,6 +1,7 @@
 import { KIcon } from '../../../KIcon';
 import type { KButtonIconProps } from './types';
 
+// Used to have same leftIcon and rightIcon props and visuals
 export function KButtonIcon(props?: KButtonIconProps) {
   if (!props?.icon) return <></>;