+++ uikit/src/components/KButton/constants.ts
@@ -2,5 +2,5 @@ import type { KButtonSizes } from './types';
 
 export const KBUTTON_ICON_SIZES: Record<KButtonSizes, number> = {
   DEFAULT: 16,
-  MINI: 8,
+  MINI: 14,
 };