+++ uikit/src/components/KButtonList/index.ts
@@ -1,0 +1,2 @@
+export type { KButtonListProps } from './types';
+export { KButtonList } from './KButtonList';