+++ uikit/src/components/KButtonList/types.ts
@@ -1,0 +1,11 @@
+import type { KButtonProps } from '../KButton/types';
+import type { KIconNames } from '../KIcon/iconNames';
+
+export type KButtonListItem = Pick<KButtonProps, 'onClick' | 'title'> & {
+  name: KIconNames;
+};
+
+export type KButtonListProps = Pick<KButtonProps, 'size'> & {
+  items?: KButtonListItem[];
+  style?: React.CSSProperties;
+};