+++ uikit/src/components/KIcon/__stories__/KIcon.stories.tsx
@@ -12,6 +12,6 @@ type Story = StoryObj<typeof KIcon>;
 export const Primary: Story = {
   args: {
     name: 'close',
-    size: 64,
+    size: 16,
   },
 };