+++ uikit/src/components/KIcon/__stories__/KIconStoriesRow.tsx
@@ -7,11 +7,23 @@ export type KIconStoriesRowProps = {
 
 export function KIconStoriesRow(props: KIconStoriesRowProps) {
   return (
-    <div className="krow" style={{ textAlign: 'center' }}>
+    <div
+      className="krow"
+      style={
+        {
+          marginTop: 30,
+          textAlign: 'center',
+          '--column-gutter': '8px',
+        } as React.CSSProperties
+      }
+    >
       {props.items.map((i) => (
-        <div className="kcol-1" style={{ textAlign: 'center' }}>
+        <div
+          className="kcol-1"
+          style={{ textAlign: 'center', fontSize: 10, lineHeight: '1.3em' }}
+        >
           <div>
-            <KIcon name={i} size={64} />
+            <KIcon name={i} size={16} />
           </div>
           <div>{i}</div>
         </div>