+++ uikit/src/components/KTitle/__tests__/helpers.tsx
@@ -1,0 +1,4 @@
+import * as KTitleStories from '../__stories__/KTitle.stories';
+import { composeStories } from '@storybook/react-vite';
+
+export const KTitleComposedStories = composeStories(KTitleStories);